requirements.txt
requests
rich
questionary
pyfiglet
google-generativeai
python-dotenv
